Move zenml.steps.step_context.py -> zenml.new.steps.step_context.py
