+++ examples/label_studio_annotation/steps/pytorch_trainer.py
@@ -32,8 +32,7 @@ from zenml.integrations.label_studio.label_studio_utils import (
 )
 from zenml.io import fileio
 from zenml.post_execution import get_pipeline
-from zenml.steps import BaseParameters, step
-from zenml.steps.step_context import StepContext
+from zenml.steps import BaseParameters, StepContext, step
 from zenml.utils import io_utils
 
 LABEL_MAPPING = {label: idx for idx, label in enumerate(LABELS)}
