+++ src/zenml/integrations/bentoml/steps/bento_builder.py
@@ -24,7 +24,7 @@ from zenml import step
 from zenml.integrations.bentoml.constants import DEFAULT_BENTO_FILENAME
 from zenml.logger import get_logger
 from zenml.materializers import UnmaterializedArtifact
-from zenml.steps.step_context import StepContext
+from zenml.steps import StepContext
 from zenml.utils import source_utils
 from zenml.utils.artifact_utils import load_artifact
